pyproject.toml
[project]
name = "inwx-cli"
version = "0.1.0"
description = "INWX API Command Line Interface"
requires-python = ">=3.11"
dependencies = [
    "inwx-domrobot",
]

[project.scripts]
inwx-cli = "inwx_cli.cli:main"

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

